rssproc/requirements.txt
Django==1.6
django-tinymce==2.2.0
feedparser==5.2.1
fuzzywuzzy==0.8.0
mysqlclient==1.3.7
nltk==3.1
python-Levenshtein==0.12.0
textblob==0.11.0
wheel==0.26.0